+++ src/views/tools/refresh/index.vue
@@ -2,10 +2,8 @@
 import { NIcon, NPopover, NButton } from 'naive-ui';
 import { ArrowClockwise16Regular } from '@vicons/fluent';
 import { useWebCutHistory } from '../../../hooks/history';
-import { useWebCutContext } from '../../../hooks';
 import { useT } from '../../../i18n/hooks';
 
-const { loading } = useWebCutContext();
 const { refreshRender } = useWebCutHistory();
 const t = useT();
 
@@ -17,7 +15,7 @@ async function handleRefresh() {
 <template>
     <n-popover :delay="200" class="webcut-tooltip">
         <template #trigger>
-            <n-button quaternary :focusable="false" @click="handleRefresh" class="webcut-tool-button" :loading="loading">
+            <n-button quaternary :focusable="false" @click="handleRefresh" class="webcut-tool-button">
                 <template #icon>
                     <n-icon :component="ArrowClockwise16Regular" size="16px"></n-icon>
                 </template>